requirements.txt.txt
python-dotenv>=1.0.1
requests>=2.32.3

langchain>=0.2.16
langchain-core>=0.2.38
langchain-community>=0.2.12
langgraph>=0.2.34

openai>=1.40.0
langchain-openai>=0.1.23

pinecone-client>=3.2.2
langchain-pinecone>=0.1.3

langchain-text-splitters>=0.2.2
